updated_chatbot/requirements.txt
streamlit
PyPDF2
langchain
langchain-community
flask
python-dotenv
chromadb
openai
tiktoken
streamlit_chat
